pyproject.toml
[tool.poetry]
name = "xilos"
version = "0.1.0"
description = "Production-grade ML project template"
authors = ["Milosz Bertman <milosz@xomnia.com>"]
readme = "README.md"
packages = [{ include = "xilos", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11.9"

numpy = "^1.26.0"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
loguru = "^0.7.2"
polars = "^0.20.0"
scipy = "^1.11.0"
fastparquet = "^2025.12.0"

[tool.poetry.group.serve.dependencies]
fastapi = "^0.109.0"
uvicorn = "^0.27.0"

[tool.poetry.group.aws.dependencies]
boto3 = "^1.34.0"

[tool.poetry.group.azure.dependencies]
azure-storage-blob = "^12.19.0"
azure-cosmos = "^4.5.0"

[tool.poetry.group.gcp.dependencies]
google-cloud-storage = "^2.14.0"
google-cloud-bigquery = "^3.17.0"
db-dtypes = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
ruff = "^0.3.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"
exclude = ["notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
